panos-log-collector/requirements.txt
black==22.1.0
blinker==1.8.2
blue==0.9.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
flake8==4.0.1
Flask==3.0.3
Flask-WTF==1.2.1
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.6.1
mypy-extensions==1.0.0
packaging==24.0
pan-os-python==1.12.1
pan-python==0.17.0
pathspec==0.12.1
platformdirs==4.2.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
tomli==2.0.1
urllib3==2.2.1
Werkzeug==3.0.3
WTForm==1.0
WTForms==3.1.2
xmltodict==0.13.0
